Add strengthify bar after plain text input clone of password input
To show the password in plain text "showPassword" adds a text input
after the password input and swaps their visibility depending on whether
the password has to be shown in plain text or not. In a similar way,
"strengthify" by default adds the strength bar after the input element
it was called on. Due to this, if "showPassword" is called before
"strengthify" on the same password input then the strength bar ends
between the password input and the text input, and when the text input
is shown it appears below the strength bar.

To fix this now the strength bar is added after the text input in those
places in which "strengthify" was called after "showPassword" on the
same element.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>