Fix "New" menu shown behind the multiselect header in the file list
The "New" menu is a descendant of the controls bar, and the controls bar
and the multiselect header belong to the same stacking context. As they
had the same z-index but the multiselect header appears after the
controls bar in the DOM the controls bar and its descendants, including
the "New" menu, were rendered behind the multiselect header. Now the
controls bar has a z-index value higher than the one used for the
multiselect header to ensure that the "New" menu is rendered in front of
the multiselect header.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>