Fix missing clipboard icon in shared links
The clipboard icon in shared links appears either directly on the link
input field or, if any social sharing app is enabled, in a menu. The
clipboard icon uses the same CSS rules as other icons (like the
information icon) to be posioned on the end of the input field, and
those rules have to be "cancelled" when the icon is shown in the menu.

Fixes #7990

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>