Allow ldap_dynamic_group_member_url to be a filter
This change allows dynamic_group_member_url of the user_ldap module to
be the filter itself instead of an attribute where the filter can
be read from. The value is assumed to be a filter when it starts with
a round bracket. The substring '%gid' will be replaced by the group's
DN.

This allows group members to be resolved by LDAP_MATCHING_RULE_IN_CHAIN
on Active Directory servers. Example:

  (&(objectClass=person)(memberOf:1.2.840.113556.1.4.1941:=%gid))

Returns all persons that are either direct members of %gid or members
of another group that is member of %gid.

Signed-off-by: Roland Tapken <roland@bitarbeiter.net>