Adjust timeouts in the step to create a new comment
Depending on the previous steps the new comment field may be already
shown or not when the step to create a new comment is executed.
Therefore, the timeout was increased from 2 to the "standard" 10 seconds
used in other tests.

If the new comment field was found there is no need to use a timeout
when looking for the new comment button; it is either there or not, it
will not appear after some time.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>