Look up nextcloud instance based off SRV
This change allows a federated user/share to be added using a hostname
different to the real hostname of the nextcloud server.

This is useful in cases where a cloud is on a different domain to an
email address as it will allow users to share directly with the email
address and have it resolve to the nextcloud user.

For example, if somebody has the email address 'me@example.com', and
a nextcloud instance at 'cloud.example.com', this change will resolve
the federated user to 'me@cloud.example.com'.

This requires the username to be the same as the email address prefix.
It also requires an appropriate _nextcloud._tcp SRV record.

Related issue: #365

Signed-off-by: Joshua Hesketh <josh@nitrotech.org>